[Pager] Consume post flings and scrolls from nested scrolling content (#455)
* [Pager] Consume post flings

Currently any nested unused scrolls + flings from child content
is handled by the Pager. This is different to how ViewPager[2]
behaves and is usually unwanted behavior.

This PR adds a NestedScrollConnection which consumes those events
before the Pager gets them. It's not perfect, but it's close enough
for now.

Fixes #347

* Add tests for scrolling content in Pager